third_party/blink/web_tests/external/wpt/css/css-lists/parsing/content-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['content'] = "counter(foo, none)" should not set the property value
  assert_equals: expected "" but got "counter(foo, none)"
[FAIL] e.style['content'] = "counters(foo, \\"\\", none)" should not set the property value
  assert_equals: expected "" but got "counter(foo, none)"
[FAIL] e.style['content'] = "\\"z\\" counter(foo, none)" should not set the property value
  assert_equals: expected "" but got "\\"z\\" counter(foo, none)"
[FAIL] e.style['content'] = "\\"z\\" counters(foo, \\"\\", none)" should not set the property value
  assert_equals: expected "" but got "\\"z\\" counter(foo, none)"
[FAIL] e.style['content'] = "counter(foo, none) \\"z\\"" should not set the property value
  assert_equals: expected "" but got "counter(foo, none) \\"z\\""
[FAIL] e.style['content'] = "counters(foo, \\"\\", none) \\"z\\"" should not set the property value
  assert_equals: expected "" but got "counter(foo, none) \\"z\\""
Harness: the test ran to completion.
